encapsulate rawclient to its own file
